7 Wonder Of The World Recognition System
This is the code of CNN implemented on a set of dataset of the 7 wonders of the world. The dataset is self made and can be manipulated as needed.
There are 2 files
1. Wonder Of The World Identifier.py

This file shows the contrustion, training and findinf the efficiency of the model and how to save the trained model, so that you don;t have to train the model again.

2. LoadTheModelAndUseIt.py

This file contain the code on how to call the file loaded with the trained model and test it on images as per your choice.